Asynchronous cache: Refactoring
