+++ src/common/base_cache.rs
@@ -1,28 +1,317 @@
-use crate::sync::cache::{
-    READ_LOG_FLUSH_POINT, WRITE_LOG_FLUSH_POINT, WRITE_LOG_LOW_WATER_MARK, WRITE_LOG_SIZE,
-};
-
 use super::{
     deque::{CacheRegion, DeqNode, Deque},
     deques::Deques,
     frequency_sketch::FrequencySketch,
-    housekeeper::{InnerSync, SyncPace},
+    housekeeper::{Housekeeper, InnerSync, SyncPace},
     AccessTime, KeyDate, KeyHash, KeyHashDate, ReadOp, ValueEntry, WriteOp,
 };
 
-use crossbeam_channel::Receiver;
+use crossbeam_channel::{Receiver, Sender, TrySendError};
 use parking_lot::{Mutex, MutexGuard, RwLock};
 use quanta::{Clock, Instant};
 use std::{
     borrow::Borrow,
+    collections::hash_map::RandomState,
     hash::{BuildHasher, Hash, Hasher},
     ptr::NonNull,
+    rc::Rc,
     sync::{
-        atomic::{AtomicBool, Ordering},
+        atomic::{AtomicBool, AtomicU8, Ordering},
         Arc,
     },
     time::Duration,
 };
+
+pub(crate) const MAX_SYNC_REPEATS: usize = 4;
+
+const READ_LOG_FLUSH_POINT: usize = 512;
+const READ_LOG_SIZE: usize = READ_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS + 2);
+
+const WRITE_LOG_FLUSH_POINT: usize = 512;
+const WRITE_LOG_LOW_WATER_MARK: usize = WRITE_LOG_FLUSH_POINT / 2;
+pub(crate) const WRITE_LOG_HIGH_WATER_MARK: usize = WRITE_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS - 1);
+const WRITE_LOG_SIZE: usize = WRITE_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS + 2);
+
+pub(crate) const WRITE_THROTTLE_MICROS: u64 = 15;
+pub(crate) const WRITE_RETRY_INTERVAL_MICROS: u64 = 50;
+
+pub(crate) const PERIODICAL_SYNC_INITIAL_DELAY_MILLIS: u64 = 500;
+pub(crate) const PERIODICAL_SYNC_NORMAL_PACE_MILLIS: u64 = 300;
+pub(crate) const PERIODICAL_SYNC_FAST_PACE_NANOS: u64 = 500;
+
+pub(crate) type HouseKeeperArc<K, V, S> = Arc<Housekeeper<Inner<K, V, S>>>;
+
+pub(crate) struct BaseCache<K, V, S = RandomState> {
+    pub(crate) inner: Arc<Inner<K, V, S>>,
+    read_op_ch: Sender<ReadOp<K, V>>,
+    pub(crate) write_op_ch: Sender<WriteOp<K, V>>,
+    pub(crate) housekeeper: Option<HouseKeeperArc<K, V, S>>,
+}
+
+impl<K, V, S> Clone for BaseCache<K, V, S> {
+    /// Makes a clone of this shared cache.
+    ///
+    /// This operation is cheap as it only creates thread-safe reference counted
+    /// pointers to the shared internal data structures.
+    fn clone(&self) -> Self {
+        Self {
+            inner: Arc::clone(&self.inner),
+            read_op_ch: self.read_op_ch.clone(),
+            write_op_ch: self.write_op_ch.clone(),
+            housekeeper: self.housekeeper.as_ref().map(|h| Arc::clone(&h)),
+        }
+    }
+}
+
+impl<K, V, S> Drop for BaseCache<K, V, S> {
+    fn drop(&mut self) {
+        // The housekeeper needs to be dropped before the inner is dropped.
+        std::mem::drop(self.housekeeper.take());
+    }
+}
+
+impl<K, V, S> BaseCache<K, V, S>
+where
+    K: Hash + Eq,
+    V: Clone,
+    S: BuildHasher + Clone,
+{
+    pub(crate) fn with_everything(
+        max_capacity: usize,
+        initial_capacity: Option<usize>,
+        build_hasher: S,
+        time_to_live: Option<Duration>,
+        time_to_idle: Option<Duration>,
+    ) -> Self {
+        let (r_snd, r_rcv) = crossbeam_channel::bounded(READ_LOG_SIZE);
+        let (w_snd, w_rcv) = crossbeam_channel::bounded(WRITE_LOG_SIZE);
+        let inner = Arc::new(Inner::new(
+            max_capacity,
+            initial_capacity,
+            build_hasher,
+            r_rcv,
+            w_rcv,
+            time_to_live,
+            time_to_idle,
+        ));
+        let housekeeper = Housekeeper::new(Arc::downgrade(&inner));
+
+        Self {
+            inner,
+            read_op_ch: r_snd,
+            write_op_ch: w_snd,
+            housekeeper: Some(Arc::new(housekeeper)),
+        }
+    }
+
+    pub(crate) fn get_with_hash<Q>(&self, key: &Q, hash: u64) -> Option<V>
+    where
+        Arc<K>: Borrow<Q>,
+        Q: Hash + Eq + ?Sized,
+    {
+        let record = |entry, ts| {
+            self.record_read_op(hash, entry, ts)
+                .expect("Failed to record a get op");
+        };
+
+        match (self.inner.get(key), self.inner.has_expiry()) {
+            // Value not found.
+            (None, _) => {
+                record(None, None);
+                None
+            }
+            // Value found, no expiry.
+            (Some(entry), false) => {
+                let v = entry.value.clone();
+                record(Some(entry), None);
+                Some(v)
+            }
+            // Value found, need to check if expired.
+            (Some(entry), true) => {
+                let now = self.inner.current_time_from_expiration_clock();
+                if self.inner.is_expired_entry_wo(&entry, now)
+                    || self.inner.is_expired_entry_ao(&entry, now)
+                {
+                    // Expired entry. Record this access as a cache miss rather than a hit.
+                    record(None, None);
+                    None
+                } else {
+                    // Valid entry.
+                    let v = entry.value.clone();
+                    record(Some(entry), Some(now));
+                    Some(v)
+                }
+            }
+        }
+    }
+
+    #[inline]
+    fn record_read_op(
+        &self,
+        hash: u64,
+        entry: Option<Arc<ValueEntry<K, V>>>,
+        timestamp: Option<Instant>,
+    ) -> Result<(), TrySendError<ReadOp<K, V>>> {
+        use ReadOp::*;
+        self.apply_reads_if_needed();
+        let ch = &self.read_op_ch;
+        let op = if let Some(entry) = entry {
+            Hit(hash, entry, timestamp)
+        } else {
+            Miss(hash)
+        };
+        match ch.try_send(op) {
+            // Discard the ReadOp when the channel is full.
+            Ok(()) | Err(TrySendError::Full(_)) => Ok(()),
+            Err(e @ TrySendError::Disconnected(_)) => Err(e),
+        }
+    }
+
+    #[inline]
+    pub(crate) fn do_insert_with_hash(&self, key: K, hash: u64, value: V) -> WriteOp<K, V> {
+        let key = Arc::new(key);
+
+        let op_cnt1 = Rc::new(AtomicU8::new(0));
+        let op_cnt2 = Rc::clone(&op_cnt1);
+        let mut op1 = None;
+        let mut op2 = None;
+
+        // Since the cache (cht::SegmentedHashMap) employs optimistic locking
+        // strategy, insert_with_or_modify() may get an insert/modify operation
+        // conflicted with other concurrent hash table operations. In that case, it
+        // has to retry the insertion or modification, so on_insert and/or on_modify
+        // closures can be executed more than once. In order to identify the last
+        // call of these closures, we use a shared counter (op_cnt{1,2}) here to
+        // record a serial number on a WriteOp, and consider the WriteOp with the
+        // largest serial number is the one made by the last call of the closures.
+        self.inner.cache.insert_with_or_modify(
+            Arc::clone(&key),
+            // on_insert
+            || {
+                let mut last_accessed = None;
+                let mut last_modified = None;
+                if self.inner.has_expiry() {
+                    let ts = unsafe { std::mem::transmute(std::u64::MAX) };
+                    if self.inner.time_to_idle.is_some() {
+                        last_accessed = Some(ts);
+                    }
+                    if self.inner.time_to_live.is_some() {
+                        last_modified = Some(ts);
+                    }
+                }
+                let entry = Arc::new(ValueEntry::new(
+                    value.clone(),
+                    last_accessed,
+                    last_modified,
+                    None,
+                    None,
+                ));
+                let cnt = op_cnt1.fetch_add(1, Ordering::Relaxed);
+                op1 = Some((
+                    cnt,
+                    WriteOp::Insert(KeyHash::new(key, hash), Arc::clone(&entry)),
+                ));
+                entry
+            },
+            // on_modify
+            |_k, old_entry| {
+                let entry = Arc::new(ValueEntry::new_with(value.clone(), old_entry));
+                let cnt = op_cnt2.fetch_add(1, Ordering::Relaxed);
+                op2 = Some((
+                    cnt,
+                    Arc::clone(&old_entry),
+                    WriteOp::Update(Arc::clone(&entry)),
+                ));
+                entry
+            },
+        );
+
+        match (op1, op2) {
+            (Some((_cnt, ins_op)), None) => ins_op,
+            (None, Some((_cnt, old_entry, upd_op))) => {
+                old_entry.unset_q_nodes();
+                upd_op
+            }
+            (Some((cnt1, ins_op)), Some((cnt2, old_entry, upd_op))) => {
+                if cnt1 > cnt2 {
+                    ins_op
+                } else {
+                    old_entry.unset_q_nodes();
+                    upd_op
+                }
+            }
+            (None, None) => unreachable!(),
+        }
+    }
+
+    #[inline]
+    fn apply_reads_if_needed(&self) {
+        let len = self.read_op_ch.len();
+
+        if Self::should_apply_reads(len) {
+            if let Some(h) = &self.housekeeper {
+                h.try_schedule_sync();
+            }
+        }
+    }
+
+    #[inline]
+    pub(crate) fn apply_reads_writes_if_needed(
+        ch: &Sender<WriteOp<K, V>>,
+        housekeeper: Option<&HouseKeeperArc<K, V, S>>,
+    ) {
+        let w_len = ch.len();
+
+        if Self::should_apply_writes(w_len) {
+            if let Some(h) = housekeeper {
+                h.try_schedule_sync();
+            }
+        }
+    }
+
+    #[inline]
+    fn should_apply_reads(ch_len: usize) -> bool {
+        ch_len >= READ_LOG_FLUSH_POINT
+    }
+
+    #[inline]
+    fn should_apply_writes(ch_len: usize) -> bool {
+        ch_len >= WRITE_LOG_FLUSH_POINT
+    }
+}
+
+#[cfg(test)]
+impl<K, V, S> BaseCache<K, V, S>
+where
+    K: Hash + Eq,
+    S: BuildHasher + Clone,
+{
+    pub(crate) fn reconfigure_for_testing(&mut self) {
+        // Stop the housekeeping job that may cause sync() method to return earlier.
+        if let Some(housekeeper) = &self.housekeeper {
+            // TODO: Extract this into a housekeeper method.
+            let mut job = housekeeper.periodical_sync_job().lock();
+            if let Some(job) = job.take() {
+                job.cancel();
+            }
+        }
+    }
+
+    pub(crate) fn set_expiration_clock(&self, clock: Option<quanta::Clock>) {
+        let mut exp_clock = self.inner.expiration_clock.write();
+        if let Some(clock) = clock {
+            *exp_clock = Some(clock);
+            self.inner
+                .has_expiration_clock
+                .store(true, Ordering::SeqCst);
+        } else {
+            self.inner
+                .has_expiration_clock
+                .store(false, Ordering::SeqCst);
+            *exp_clock = None;
+        }
+    }
+}
 
 type CacheStore<K, V, S> = cht::SegmentedHashMap<Arc<K>, Arc<ValueEntry<K, V>>, S>;
