+++ src/common/housekeeper.rs
@@ -2,7 +2,7 @@ use super::{
     thread_pool::{ThreadPool, ThreadPoolRegistry},
     unsafe_weak_pointer::UnsafeWeakPointer,
 };
-use crate::sync::cache::{
+use crate::common::base_cache::{
     MAX_SYNC_REPEATS, PERIODICAL_SYNC_FAST_PACE_NANOS, PERIODICAL_SYNC_INITIAL_DELAY_MILLIS,
     PERIODICAL_SYNC_NORMAL_PACE_MILLIS,
 };
