+++ src/sync/cache.rs
@@ -1,40 +1,21 @@
 use super::ConcurrentCacheExt;
 use crate::common::{
-    base_cache::Inner,
-    housekeeper::{Housekeeper, InnerSync},
-    KeyHash, ReadOp, ValueEntry, WriteOp,
+    base_cache::{
+        BaseCache, HouseKeeperArc, MAX_SYNC_REPEATS, WRITE_LOG_HIGH_WATER_MARK,
+        WRITE_RETRY_INTERVAL_MICROS, WRITE_THROTTLE_MICROS,
+    },
+    housekeeper::InnerSync,
+    WriteOp,
 };
 
 use crossbeam_channel::{Sender, TrySendError};
-use quanta::Instant;
 use std::{
     borrow::Borrow,
     collections::hash_map::RandomState,
     hash::{BuildHasher, Hash},
-    rc::Rc,
-    sync::{
-        atomic::{AtomicU8, Ordering},
-        Arc,
-    },
+    sync::Arc,
     time::Duration,
 };
-
-pub(crate) const MAX_SYNC_REPEATS: usize = 4;
-
-pub(crate) const READ_LOG_FLUSH_POINT: usize = 512;
-pub(crate) const READ_LOG_SIZE: usize = READ_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS + 2);
-
-pub(crate) const WRITE_LOG_FLUSH_POINT: usize = 512;
-pub(crate) const WRITE_LOG_LOW_WATER_MARK: usize = WRITE_LOG_FLUSH_POINT / 2;
-pub(crate) const WRITE_LOG_HIGH_WATER_MARK: usize = WRITE_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS - 1);
-pub(crate) const WRITE_LOG_SIZE: usize = WRITE_LOG_FLUSH_POINT * (MAX_SYNC_REPEATS + 2);
-
-pub(crate) const WRITE_THROTTLE_MICROS: u64 = 15;
-pub(crate) const WRITE_RETRY_INTERVAL_MICROS: u64 = 50;
-
-pub(crate) const PERIODICAL_SYNC_INITIAL_DELAY_MILLIS: u64 = 500;
-pub(crate) const PERIODICAL_SYNC_NORMAL_PACE_MILLIS: u64 = 300;
-pub(crate) const PERIODICAL_SYNC_FAST_PACE_NANOS: u64 = 500;
 
 /// A thread-safe concurrent in-memory cache.
 ///
@@ -192,18 +173,9 @@ pub(crate) const PERIODICAL_SYNC_FAST_PACE_NANOS: u64 = 500;
 /// [build-with-hasher-method]: ./struct.CacheBuilder.html#method.build_with_hasher
 /// [ahash-crate]: https://crates.io/crates/ahash
 ///
+#[derive(Clone)]
 pub struct Cache<K, V, S = RandomState> {
-    inner: Arc<Inner<K, V, S>>,
-    read_op_ch: Sender<ReadOp<K, V>>,
-    write_op_ch: Sender<WriteOp<K, V>>,
-    housekeeper: Option<Arc<Housekeeper<Inner<K, V, S>>>>,
-}
-
-impl<K, V, S> Drop for Cache<K, V, S> {
-    fn drop(&mut self) {
-        // The housekeeper needs to be dropped before the inner is dropped.
-        std::mem::drop(self.housekeeper.take());
-    }
+    base: BaseCache<K, V, S>,
 }
 
 unsafe impl<K, V, S> Send for Cache<K, V, S>
@@ -220,21 +192,6 @@ where
     V: Send + Sync,
     S: Sync,
 {
-}
-
-impl<K, V, S> Clone for Cache<K, V, S> {
-    /// Makes a clone of this shared cache.
-    ///
-    /// This operation is cheap as it only creates thread-safe reference counted
-    /// pointers to the shared internal data structures.
-    fn clone(&self) -> Self {
-        Self {
-            inner: Arc::clone(&self.inner),
-            read_op_ch: self.read_op_ch.clone(),
-            write_op_ch: self.write_op_ch.clone(),
-            housekeeper: self.housekeeper.as_ref().map(|h| Arc::clone(&h)),
-        }
-    }
 }
 
 impl<K, V> Cache<K, V, RandomState>
@@ -267,24 +224,14 @@ where
         time_to_live: Option<Duration>,
         time_to_idle: Option<Duration>,
     ) -> Self {
-        let (r_snd, r_rcv) = crossbeam_channel::bounded(READ_LOG_SIZE);
-        let (w_snd, w_rcv) = crossbeam_channel::bounded(WRITE_LOG_SIZE);
-        let inner = Arc::new(Inner::new(
-            max_capacity,
-            initial_capacity,
-            build_hasher,
-            r_rcv,
-            w_rcv,
-            time_to_live,
-            time_to_idle,
-        ));
-        let housekeeper = Housekeeper::new(Arc::downgrade(&inner));
-
         Self {
-            inner,
-            read_op_ch: r_snd,
-            write_op_ch: w_snd,
-            housekeeper: Some(Arc::new(housekeeper)),
+            base: BaseCache::with_everything(
+                max_capacity,
+                initial_capacity,
+                build_hasher,
+                time_to_live,
+                time_to_idle,
+            ),
         }
     }
 
@@ -303,15 +250,30 @@ where
         Arc<K>: Borrow<Q>,
         Q: Hash + Eq + ?Sized,
     {
-        self.get_with_hash(key, self.inner.hash(key))
+        self.base.get_with_hash(key, self.base.inner.hash(key))
+    }
+
+    pub(crate) fn get_with_hash<Q>(&self, key: &Q, hash: u64) -> Option<V>
+    where
+        Arc<K>: Borrow<Q>,
+        Q: Hash + Eq + ?Sized,
+    {
+        self.base.get_with_hash(key, hash)
     }
 
     /// Inserts a key-value pair into the cache.
     ///
     /// If the cache has this key present, the value is updated.
     pub fn insert(&self, key: K, value: V) {
-        let hash = self.inner.hash(&key);
+        let hash = self.base.inner.hash(&key);
         self.insert_with_hash(key, hash, value)
+    }
+
+    pub(crate) fn insert_with_hash(&self, key: K, hash: u64, value: V) {
+        self.throttle_write_pace();
+        let op = self.base.do_insert_with_hash(key, hash, value);
+        let hk = self.base.housekeeper.as_ref();
+        Self::schedule_write_op(&self.base.write_op_ch, op, hk).expect("Failed to insert");
     }
 
     /// Discards any cached value for the key.
@@ -324,24 +286,26 @@ where
         Q: Hash + Eq + ?Sized,
     {
         self.throttle_write_pace();
-        if let Some(entry) = self.inner.cache.remove(key) {
-            self.schedule_remove_op(entry).expect("Failed to remove");
+        if let Some(entry) = self.base.inner.cache.remove(key) {
+            let op = WriteOp::Remove(entry);
+            let hk = self.base.housekeeper.as_ref();
+            Self::schedule_write_op(&self.base.write_op_ch, op, hk).expect("Failed to remove");
         }
     }
 
     /// Returns the `max_capacity` of this cache.
     pub fn max_capacity(&self) -> usize {
-        self.inner.max_capacity
+        self.base.inner.max_capacity
     }
 
     /// Returns the `time_to_live` of this cache.
     pub fn time_to_live(&self) -> Option<Duration> {
-        self.inner.time_to_live
+        self.base.inner.time_to_live
     }
 
     /// Returns the `time_to_idle` of this cache.
     pub fn time_to_idle(&self) -> Option<Duration> {
-        self.inner.time_to_idle
+        self.base.inner.time_to_idle
     }
 
     /// Returns the number of internal segments of this cache.
@@ -349,126 +313,6 @@ where
     /// `Cache` always returns `1`.
     pub fn num_segments(&self) -> usize {
         1
-    }
-
-    pub(crate) fn get_with_hash<Q>(&self, key: &Q, hash: u64) -> Option<V>
-    where
-        Arc<K>: Borrow<Q>,
-        Q: Hash + Eq + ?Sized,
-    {
-        let record = |entry, ts| {
-            self.record_read_op(hash, entry, ts)
-                .expect("Failed to record a get op")
-        };
-
-        match (self.inner.get(key), self.inner.has_expiry()) {
-            // Value not found.
-            (None, _) => {
-                record(None, None);
-                None
-            }
-            // Value found, no expiry.
-            (Some(entry), false) => {
-                let v = entry.value.clone();
-                record(Some(entry), None);
-                Some(v)
-            }
-            // Value found, need to check if expired.
-            (Some(entry), true) => {
-                let now = self.inner.current_time_from_expiration_clock();
-                if self.inner.is_expired_entry_wo(&entry, now)
-                    || self.inner.is_expired_entry_ao(&entry, now)
-                {
-                    // Expired entry. Record this access as a cache miss rather than a hit.
-                    record(None, None);
-                    None
-                } else {
-                    // Valid entry.
-                    let v = entry.value.clone();
-                    record(Some(entry), Some(now));
-                    Some(v)
-                }
-            }
-        }
-    }
-
-    pub(crate) fn insert_with_hash(&self, key: K, hash: u64, value: V) {
-        self.throttle_write_pace();
-
-        let key = Arc::new(key);
-
-        let op_cnt1 = Rc::new(AtomicU8::new(0));
-        let op_cnt2 = Rc::clone(&op_cnt1);
-        let mut op1 = None;
-        let mut op2 = None;
-
-        // Since the cache (cht::SegmentedHashMap) employs optimistic locking
-        // strategy, insert_with_or_modify() may get an insert/modify operation
-        // conflicted with other concurrent hash table operations. In that case, it
-        // has to retry the insertion or modification, so on_insert and/or on_modify
-        // closures can be executed more than once. In order to identify the last
-        // call of these closures, we use a shared counter (op_cnt{1,2}) here to
-        // record a serial number on a WriteOp, and consider the WriteOp with the
-        // largest serial number is the one made by the last call of the closures.
-        self.inner.cache.insert_with_or_modify(
-            Arc::clone(&key),
-            // on_insert
-            || {
-                let mut last_accessed = None;
-                let mut last_modified = None;
-                if self.inner.has_expiry() {
-                    let ts = unsafe { std::mem::transmute(std::u64::MAX) };
-                    if self.inner.time_to_idle.is_some() {
-                        last_accessed = Some(ts);
-                    }
-                    if self.inner.time_to_live.is_some() {
-                        last_modified = Some(ts);
-                    }
-                }
-                let entry = Arc::new(ValueEntry::new(
-                    value.clone(),
-                    last_accessed,
-                    last_modified,
-                    None,
-                    None,
-                ));
-                let cnt = op_cnt1.fetch_add(1, Ordering::Relaxed);
-                op1 = Some((
-                    cnt,
-                    WriteOp::Insert(KeyHash::new(key, hash), Arc::clone(&entry)),
-                ));
-                entry
-            },
-            // on_modify
-            |_k, old_entry| {
-                let entry = Arc::new(ValueEntry::new_with(value.clone(), old_entry));
-                let cnt = op_cnt2.fetch_add(1, Ordering::Relaxed);
-                op2 = Some((
-                    cnt,
-                    Arc::clone(&old_entry),
-                    WriteOp::Update(Arc::clone(&entry)),
-                ));
-                entry
-            },
-        );
-
-        match (op1, op2) {
-            (Some((_cnt, ins_op)), None) => self.schedule_insert_op(ins_op),
-            (None, Some((_cnt, old_entry, upd_op))) => {
-                old_entry.unset_q_nodes();
-                self.schedule_insert_op(upd_op)
-            }
-            (Some((cnt1, ins_op)), Some((cnt2, old_entry, upd_op))) => {
-                if cnt1 > cnt2 {
-                    self.schedule_insert_op(ins_op)
-                } else {
-                    old_entry.unset_q_nodes();
-                    self.schedule_insert_op(upd_op)
-                }
-            }
-            (None, None) => unreachable!(),
-        }
-        .expect("Failed to insert");
     }
 }
 
@@ -478,7 +322,7 @@ where
     S: BuildHasher + Clone,
 {
     fn sync(&self) {
-        self.inner.sync(MAX_SYNC_REPEATS);
+        self.base.inner.sync(MAX_SYNC_REPEATS);
     }
 }
 
@@ -486,33 +330,15 @@ where
 impl<K, V, S> Cache<K, V, S>
 where
     K: Hash + Eq,
+    V: Clone,
     S: BuildHasher + Clone,
 {
     #[inline]
-    fn record_read_op(
-        &self,
-        hash: u64,
-        entry: Option<Arc<ValueEntry<K, V>>>,
-        timestamp: Option<Instant>,
-    ) -> Result<(), TrySendError<ReadOp<K, V>>> {
-        use ReadOp::*;
-        self.apply_reads_if_needed();
-        let ch = &self.read_op_ch;
-        let op = if let Some(entry) = entry {
-            Hit(hash, entry, timestamp)
-        } else {
-            Miss(hash)
-        };
-        match ch.try_send(op) {
-            // Discard the ReadOp when the channel is full.
-            Ok(()) | Err(TrySendError::Full(_)) => Ok(()),
-            Err(e @ TrySendError::Disconnected(_)) => Err(e),
-        }
-    }
-
-    #[inline]
-    fn schedule_insert_op(&self, op: WriteOp<K, V>) -> Result<(), TrySendError<WriteOp<K, V>>> {
-        let ch = &self.write_op_ch;
+    fn schedule_write_op(
+        ch: &Sender<WriteOp<K, V>>,
+        op: WriteOp<K, V>,
+        housekeeper: Option<&HouseKeeperArc<K, V, S>>,
+    ) -> Result<(), TrySendError<WriteOp<K, V>>> {
         let mut op = op;
 
         // NOTES:
@@ -520,7 +346,7 @@ where
         // - We are doing a busy-loop here. We were originally calling `ch.send(op)?`,
         //   but we got a notable performance degradation.
         loop {
-            self.apply_reads_writes_if_needed();
+            BaseCache::apply_reads_writes_if_needed(ch, housekeeper);
             match ch.try_send(op) {
                 Ok(()) => break,
                 Err(TrySendError::Full(op1)) => {
@@ -531,69 +357,11 @@ where
             }
         }
         Ok(())
-    }
-
-    #[inline]
-    fn schedule_remove_op(
-        &self,
-        entry: Arc<ValueEntry<K, V>>,
-    ) -> Result<(), TrySendError<WriteOp<K, V>>> {
-        let ch = &self.write_op_ch;
-        let mut op = WriteOp::Remove(entry);
-
-        // NOTES:
-        // - This will block when the channel is full.
-        // - For the reason why we are doing a busy-loop here, the comments in
-        //   `schedule_insert_op()`.
-        loop {
-            self.apply_reads_writes_if_needed();
-            match ch.try_send(op) {
-                Ok(()) => break,
-                Err(TrySendError::Full(op1)) => {
-                    op = op1;
-                    std::thread::sleep(Duration::from_micros(WRITE_RETRY_INTERVAL_MICROS));
-                }
-                Err(e @ TrySendError::Disconnected(_)) => return Err(e),
-            }
-        }
-        Ok(())
-    }
-
-    #[inline]
-    fn apply_reads_if_needed(&self) {
-        let len = self.read_op_ch.len();
-
-        if self.should_apply_reads(len) {
-            if let Some(h) = &self.housekeeper {
-                h.try_schedule_sync();
-            }
-        }
-    }
-
-    #[inline]
-    fn apply_reads_writes_if_needed(&self) {
-        let w_len = self.write_op_ch.len();
-
-        if self.should_apply_writes(w_len) {
-            if let Some(h) = &self.housekeeper {
-                h.try_schedule_sync();
-            }
-        }
-    }
-
-    #[inline]
-    fn should_apply_reads(&self, ch_len: usize) -> bool {
-        ch_len >= READ_LOG_FLUSH_POINT
-    }
-
-    #[inline]
-    fn should_apply_writes(&self, ch_len: usize) -> bool {
-        ch_len >= WRITE_LOG_FLUSH_POINT
     }
 
     #[inline]
     fn throttle_write_pace(&self) {
-        if self.write_op_ch.len() >= WRITE_LOG_HIGH_WATER_MARK {
+        if self.base.write_op_ch.len() >= WRITE_LOG_HIGH_WATER_MARK {
             std::thread::sleep(Duration::from_micros(WRITE_THROTTLE_MICROS))
         }
     }
@@ -607,29 +375,11 @@ where
     S: BuildHasher + Clone,
 {
     fn reconfigure_for_testing(&mut self) {
-        // Stop the housekeeping job that may cause sync() method to return earlier.
-        if let Some(housekeeper) = &self.housekeeper {
-            // TODO: Extract this into a housekeeper method.
-            let mut job = housekeeper.periodical_sync_job().lock();
-            if let Some(job) = job.take() {
-                job.cancel();
-            }
-        }
+        self.base.reconfigure_for_testing();
     }
 
     fn set_expiration_clock(&self, clock: Option<quanta::Clock>) {
-        let mut exp_clock = self.inner.expiration_clock.write();
-        if let Some(clock) = clock {
-            *exp_clock = Some(clock);
-            self.inner
-                .has_expiration_clock
-                .store(true, Ordering::SeqCst);
-        } else {
-            self.inner
-                .has_expiration_clock
-                .store(false, Ordering::SeqCst);
-            *exp_clock = None;
-        }
+        self.base.set_expiration_clock(clock);
     }
 }
 
@@ -745,18 +495,18 @@ mod tests {
         cache.sync();
 
         assert_eq!(cache.get(&"a"), None);
-        assert!(cache.inner.cache.is_empty());
+        assert!(cache.base.inner.cache.is_empty());
 
         cache.insert("b", "bob");
         cache.sync();
 
-        assert_eq!(cache.inner.cache.len(), 1);
+        assert_eq!(cache.base.inner.cache.len(), 1);
 
         mock.increment(Duration::from_secs(5)); // 15 secs.
         cache.sync();
 
         assert_eq!(cache.get(&"b"), Some("bob"));
-        assert_eq!(cache.inner.cache.len(), 1);
+        assert_eq!(cache.base.inner.cache.len(), 1);
 
         cache.insert("b", "bill");
         cache.sync();
@@ -765,14 +515,14 @@ mod tests {
         cache.sync();
 
         assert_eq!(cache.get(&"b"), Some("bill"));
-        assert_eq!(cache.inner.cache.len(), 1);
+        assert_eq!(cache.base.inner.cache.len(), 1);
 
         mock.increment(Duration::from_secs(5)); // 25 secs
         cache.sync();
 
         assert_eq!(cache.get(&"a"), None);
         assert_eq!(cache.get(&"b"), None);
-        assert!(cache.inner.cache.is_empty());
+        assert!(cache.base.inner.cache.is_empty());
     }
 
     #[test]
@@ -803,20 +553,20 @@ mod tests {
         cache.insert("b", "bob");
         cache.sync();
 
-        assert_eq!(cache.inner.cache.len(), 2);
+        assert_eq!(cache.base.inner.cache.len(), 2);
 
         mock.increment(Duration::from_secs(5)); // 15 secs.
         cache.sync();
 
         assert_eq!(cache.get(&"a"), None);
         assert_eq!(cache.get(&"b"), Some("bob"));
-        assert_eq!(cache.inner.cache.len(), 1);
+        assert_eq!(cache.base.inner.cache.len(), 1);
 
         mock.increment(Duration::from_secs(10)); // 25 secs
         cache.sync();
 
         assert_eq!(cache.get(&"a"), None);
         assert_eq!(cache.get(&"b"), None);
-        assert!(cache.inner.cache.is_empty());
+        assert!(cache.base.inner.cache.is_empty());
     }
 }
